Let a transaction save from the amount alone and store money exactly.
Also add a confirmed wipe of all transactions, use the system category control, and ship an opaque home-screen icon.

+++ ExpenseTracer/Model/ChartModel.swift
@@ -11,12 +11,12 @@ struct ChartGroup: Identifiable, Sendable {
     let id: UUID = .init()
     var date: Date
     var categories: [ChartCategory]
-    var totalIncome: Double
-    var totalExpense: Double
+    var totalIncome: Decimal
+    var totalExpense: Decimal
 }
 
 struct ChartCategory: Identifiable, Sendable {
     let id: UUID = .init()
-    var totalValue: Double
+    var totalValue: Decimal
     var category: CategoryItem
 }
